pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "huetuber"
version = "0.1.0"
description = "GPU-accelerated stain normalization for histopathology images"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Yimin Zheng", email = "yzheng@cemm.at"},
]
dependencies = [
    "cucim>=23.10.0",
    "cupy-cuda12x>=13.6.0",
    "pillow>=11.3.0",
    "scipy>=1.16.2",
]

[dependency-groups]
benchmark = [
    "matplotlib>=3.10.6",
    "mpl-fontkit>=0.5.1",
    "pandas>=2.3.2",
    "seaborn>=0.13.2",
    "torchstain>=1.4.1",
    "unhcrpyplotstyle>=0.1.3",
]
dev = [
    "ipykernel>=6.30.1",
    "opencv-python-headless>=4.12.0.88",
    "prek>=0.1.6",
    "pytest>=8.4.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[tool.uv]
default-groups = ["dev", "benchmark"]
